requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
jinja2
python-multipart
httpx
transformers
torch
sentence-transformers
scikit-learn
faiss-cpu
prometheus-client
python-jose[cryptography]
passlib[bcrypt]
pandas
numpy
pytest
